Added utility module
